// node/typings/affiliateNotification.ts
export interface AffiliateNotification {
  IdSku: string
  ProductId: number
  An: string
  IdAffiliate: string
  Version: string
  DateModified: string
  IsActive: boolean
  StockModified: boolean
  PriceModified: boolean
  HasStockKeepingUnitModified: boolean
  HasStockKeepingUnitRemovedFromAffiliate: boolean
}