// node/typings/broadcastNotification.ts
export interface BroadcastNotification {
  IdSku: string
  ProductId: number
  An: string
  IdAffiliate: string
  Version: string
  DateModified: string
  IsActive: boolean
  StockModified: boolean
  PriceModified: boolean
  HasStockKeepingUnitModified: boolean
  HasStockKeepingUnitRemovedFromAffiliate: boolean
}