// node/typings/shopperAddress.ts
export interface ShopperAddress {
  addressName: string
  addressType: string
  city: string
  complement: unknown
  country: string
  countryfake: unknown
  geoCoordinate: number[]
  neighborhood: unknown
  number: unknown
  postalCode: string
  receiverName: string
  reference: unknown
  state: string
  street: string
  userId: string
  id: string
  accountId: string
  accountName: string
  dataEntityId: string
  createdBy: string
  createdIn: string
  updatedBy: string | null
  updatedIn: string | null
  lastInteractionBy: string
  lastInteractionIn: string
  followers: unknown[]
  tags: unknown[]
  auto_filter: unknown
}